refactor: 파일명 Exhibition으로 수정, 검색 func 이동
